# backend/doc/patient-openapi.yaml
openapi: 3.0.3
info:
  title: Patient API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local backend server
paths:
...
  /api/patients/{id}:
    get:
      tags: [Patients]
      operationId: getPatientCard
      summary: Get patient card
      description: >
        Returns a patient card in FULL or ANONYMIZED mode depending on the actor role and region access.
        Patients can access only their own card.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Patient card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientCardResponse'
        '401':
          description: Authentication required
        '403':
          description: Patient is trying to access a foreign card
        '404':
          description: Patient card not found
    patch:
      tags: [Patients]
      operationId: updatePatientCard
      summary: Edit patient card
      description: >
        Updates the patient card.
        A regular doctor can edit only patients from the doctor's own region.
        A doctor with extended permissions can edit any patient in the system.
        If a regular doctor tries to edit a patient from another region, the API returns 403.
        The doctor can update all patient card fields and the patient password, but cannot change the patient login.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientRequest'
      responses:
        '200':
          description: Patient card updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientCardResponse'
        '400':
          description: Validation failed
        '401':
          description: Authentication required
        '403':
          description: Access denied for patient card update
        '404':
          description: Patient card not found
  /api/patients/{id}/examinations:
    post:
      tags: [Examinations]
      operationId: addPatientExamination
      summary: Add patient examination record
      description: >
        Creates a new examination record for the patient with the measurement date and a set of measured values.
        Existing records are not overwritten. The examination history remains append-only.
        A regular doctor can add an examination only for a patient from the doctor's own region.
        A doctor with extended permissions can add an examination for any patient.
        Duplicate characteristicCode values inside one request are rejected.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExaminationRequest'
      responses:
        '201':
          description: Examination record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExaminationResponse'
        '400':
          description: Validation failed
        '401':
          description: Authentication required
        '403':
          description: Access denied for patient examination update
        '404':
          description: Patient card or characteristic was not found
    get:
      tags: [Examinations]
      operationId: listPatientExaminations
      summary: Get patient examination records
      description: >
        Returns the full examination history for the patient in chronological order.
        Each examination record is stored separately and the history is not overwritten.
        Patients can access only their own examination history.
        Doctors can read examination history for patients from any region.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Examination history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExaminationListResponse'
        '401':
          description: Authentication required
        '403':
          description: Access denied for patient examination history
        '404':
          description: Patient card not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateValveRequest:
      type: object
      properties:
        name:
          type: string
        size:
          type: string
        material:
          type: string
      required: [name, size, material]
    CreateOperationParametersRequest:
      type: object
      properties:
        anesthesia:
          type: string
        durationMinutes:
          type: integer
        deliverySystem:
          type: string
      required: [anesthesia, durationMinutes, deliverySystem]
...
    UpdatePatientRequest:
      type: object
      properties:
        lastName:
          type: string
        firstName:
          type: string
        middleName:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date
        diagnosis:
          type: string
        regionId:
          type: integer
          format: int64
        medications:
          type: string
        valve:
          $ref: '#/components/schemas/CreateValveRequest'
        operationParameters:
          $ref: '#/components/schemas/CreateOperationParametersRequest'
        password:
          type: string
          minLength: 6
    CreateExaminationMeasurementRequest:
      type: object
      properties:
        characteristicCode:
          type: string
        value:
          type: number
        comment:
          type: string
          nullable: true
      required: [characteristicCode, value]
    CreateExaminationRequest:
      type: object
      properties:
        title:
          type: string
        examDate:
          type: string
          format: date
        comment:
          type: string
          nullable: true
        measurements:
          type: array
          items:
            $ref: '#/components/schemas/CreateExaminationMeasurementRequest'
      required: [title, examDate, measurements]
    ValveResponse:
      type: object
      properties:
        name:
          type: string
        size:
          type: string
        material:
          type: string
      required: [name, size, material]
    OperationParametersResponse:
      type: object
      properties:
        anesthesia:
          type: string
        durationMinutes:
          type: integer
        deliverySystem:
          type: string
      required: [anesthesia, durationMinutes, deliverySystem]
...
    MeasurementResponse:
      type: object
      properties:
        characteristicId:
          type: integer
          format: int64
        characteristicCode:
          type: string
        characteristicName:
          type: string
        value:
          type: string
        unit:
          type: string
        comment:
          type: string
          nullable: true
      required: [characteristicId, characteristicCode, characteristicName, value, unit]
    VitalsHistoryItemResponse:
      type: object
      properties:
        examId:
          type: integer
          format: int64
        title:
          type: string
        examDate:
          type: string
          format: date
        comment:
          type: string
          nullable: true
        measurements:
          type: array
          items:
            $ref: '#/components/schemas/MeasurementResponse'
      required: [examId, title, examDate, measurements]
    ExaminationResponse:
      type: object
      properties:
        examId:
          type: integer
          format: int64
        title:
          type: string
        examDate:
          type: string
          format: date
        comment:
          type: string
          nullable: true
        measurements:
          type: array
          items:
            $ref: '#/components/schemas/MeasurementResponse'
      required: [examId, title, examDate, measurements]
    ExaminationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExaminationResponse'
      required: [items]
    PatientCardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        viewMode:
          type: string
          enum: [FULL, ANONYMIZED]
        patientCode:
          type: string
        lastName:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        middleName:
          type: string
          nullable: true
        regionId:
          type: integer
          format: int64
        regionName:
          type: string
        birthDate:
          type: string
          format: date
        diagnosis:
          type: string
        valve:
          $ref: '#/components/schemas/ValveResponse'
        operationParameters:
          $ref: '#/components/schemas/OperationParametersResponse'
        medications:
          type: string
        createdAt:
          type: string
          format: date-time
        vitalsHistory:
          type: array
          items:
            $ref: '#/components/schemas/VitalsHistoryItemResponse'
      required: [id, viewMode, patientCode, regionId, regionName, birthDate, diagnosis, valve, operationParameters, medications, createdAt, vitalsHistory]